:sparkles: added Drawer navigation. Removed screens that are yet to be implemented. Added screen shots of the actual state of the app
