+++ src/app/(drawer)/(tabs)/person/_layout.tsx
@@ -1,3 +1,4 @@
+import { DrawerToggleButton } from "@react-navigation/drawer";
 import { Stack } from "expo-router";
 import { setStatusBarStyle } from "expo-status-bar";
 
@@ -14,7 +15,19 @@ const PeopleLayout = () => {
           animation: "slide_from_right",
         }}
       >
-        <Stack.Screen name="index" options={{ title: "People" }} />
+        <Stack.Screen
+          name="index"
+          options={{
+            title: "People",
+            headerLeft: () => (
+              <DrawerToggleButton
+                tintColor="white"
+                pressOpacity={0.5}
+                pressColor="gray"
+              />
+            ),
+          }}
+        />
         <Stack.Screen name="[id]" options={{ title: "" }} />
       </Stack>
     </>
