+++ src/app/(drawer)/(tabs)/tv/index.tsx
@@ -1,11 +1,11 @@
 import { View, Text, ActivityIndicator, StyleSheet } from "react-native";
 import React, { useState } from "react";
-import { useGetSeries } from "../../../hooks/useGetSeries";
 import { FlatList, ScrollView } from "react-native-gesture-handler";
 import CommonCard from "@/components/common/CommonCard";
 import RowFilter from "@/components/common/RowFilter";
 import { Grid } from "react-native-animated-spinkit";
 import { FormattedSeries } from "@/lib/types";
+import { useGetSeries } from "@/hooks/useGetSeries";
 
 export default function TV() {
   const [selected, setSelected] = useState<string>("day");
