--- src/app/(tabs)/search.tsx
@@ -1,10 +1,0 @@
-import { View, Text } from "react-native";
-import React from "react";
-
-export default function Search() {
-  return (
-    <View>
-      <Text>Search</Text>
-    </View>
-  );
-}
