--- src/app/(tabs)/settings.tsx
@@ -1,10 +1,0 @@
-import { View, Text } from "react-native";
-import React from "react";
-
-export default function Settings() {
-  return (
-    <View>
-      <Text>settings</Text>
-    </View>
-  );
-}
